Update index.js
Errors:

The createActor function does not handle errors properly. If the agent.fetchRootKey() call fails, it only logs a warning and proceeds with the actor creation. This could lead to certificate validation errors later on.
The createActor function does not check if the canisterId is a valid canister ID. This could lead to errors if the canister ID is not valid.
Typos:

The ICP_RUST_BOILERPLATE_BACKEND_CANISTER_ID environment variable is not camel-cased correctly. It should be ICP_Rust_Boilerplate_Backend_Canister_ID.
Bugs:

The createActor function does not cache the actor instance. This means that a new actor instance is created every time the function is called. This could be inefficient if the actor is used frequently.
Here are some suggestions for fixing the errors and bugs:

Errors:

To handle the agent.fetchRootKey() error properly, you could throw an error or return an Err result from the createActor function. This would allow the caller to handle the error appropriately.
To check if the canisterId is valid, you could use the Actor.isValidCanisterId() function. If the canisterId is not valid, you could throw an error or return an Err result from the createActor function.
Typos:

You could rename the ICP_RUST_BOILERPLATE_BACKEND_CANISTER_ID environment variable to ICP_Rust_Boilerplate_Backend_Canister_ID.
Bugs:

To cache the actor instance, you could create a global variable to store the actor instance. This would allow the createActor function to return the cached actor instance if it already exists.